.history/requirements-dev_20220329153821.txt
Django~=3.1.2
dj-database-url==0.5.0
dj-static==0.0.6
python-decouple==3.3
djangorestframework